[API TEST] Rejestracja i Logowanie

+++ init-db.sql
@@ -6,7 +6,7 @@ CREATE SCHEMA IF NOT EXISTS groups;
 
 -- Utworzenie tabel w schemacie users
 CREATE TABLE IF NOT EXISTS users.users (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     username VARCHAR(100) NOT NULL UNIQUE,
     email VARCHAR(255) NOT NULL UNIQUE,
     password VARCHAR(255) NOT NULL,
@@ -15,30 +15,30 @@ CREATE TABLE IF NOT EXISTS users.users (
 );
 
 CREATE TABLE IF NOT EXISTS users.roles (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(50) NOT NULL UNIQUE
 );
 
 CREATE TABLE IF NOT EXISTS users.user_roles (
-    user_id INTEGER REFERENCES users.users(id),
-    role_id INTEGER REFERENCES users.roles(id),
+    user_id BIGINT REFERENCES users.users(id),
+    role_id BIGINT REFERENCES users.roles(id),
     PRIMARY KEY (user_id, role_id)
 );
 
 -- Utworzenie tabel w schemacie flashcards
 CREATE TABLE IF NOT EXISTS flashcards.flashcard_sets (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description TEXT,
-    owner_id INTEGER REFERENCES users.users(id),
+    owner_id BIGINT REFERENCES users.users(id),
     is_public BOOLEAN DEFAULT false,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE TABLE IF NOT EXISTS flashcards.flashcards (
-    id SERIAL PRIMARY KEY,
-    set_id INTEGER REFERENCES flashcards.flashcard_sets(id),
+    id BIGSERIAL PRIMARY KEY,
+    set_id BIGINT REFERENCES flashcards.flashcard_sets(id),
     question TEXT NOT NULL,
     answer TEXT NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -47,18 +47,18 @@ CREATE TABLE IF NOT EXISTS flashcards.flashcards (
 
 -- Utworzenie tabel w schemacie quizzes
 CREATE TABLE IF NOT EXISTS quizzes.quiz_sets (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description TEXT,
-    owner_id INTEGER REFERENCES users.users(id),
+    owner_id BIGINT REFERENCES users.users(id),
     is_public BOOLEAN DEFAULT false,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE TABLE IF NOT EXISTS quizzes.questions (
-    id SERIAL PRIMARY KEY,
-    quiz_id INTEGER REFERENCES quizzes.quiz_sets(id),
+    id BIGSERIAL PRIMARY KEY,
+    quiz_id BIGINT REFERENCES quizzes.quiz_sets(id),
     question TEXT NOT NULL,
     question_type VARCHAR(50) NOT NULL, -- multiple_choice, open, drag_drop, etc.
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -66,17 +66,17 @@ CREATE TABLE IF NOT EXISTS quizzes.questions (
 );
 
 CREATE TABLE IF NOT EXISTS quizzes.answers (
-    id SERIAL PRIMARY KEY,
-    question_id INTEGER REFERENCES quizzes.questions(id),
+    id BIGSERIAL PRIMARY KEY,
+    question_id BIGINT REFERENCES quizzes.questions(id),
     content TEXT NOT NULL,
     is_correct BOOLEAN DEFAULT false,
     points INTEGER DEFAULT 1
 );
 
 CREATE TABLE IF NOT EXISTS quizzes.results (
-    id SERIAL PRIMARY KEY,
-    user_id INTEGER REFERENCES users.users(id),
-    quiz_id INTEGER REFERENCES quizzes.quiz_sets(id),
+    id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT REFERENCES users.users(id),
+    quiz_id BIGINT REFERENCES quizzes.quiz_sets(id),
     score INTEGER NOT NULL,
     max_score INTEGER NOT NULL,
     completed_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
@@ -84,27 +84,27 @@ CREATE TABLE IF NOT EXISTS quizzes.results (
 
 -- Utworzenie tabel w schemacie groups
 CREATE TABLE IF NOT EXISTS groups.groups (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description TEXT,
-    owner_id INTEGER REFERENCES users.users(id),
+    owner_id BIGINT REFERENCES users.users(id),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE TABLE IF NOT EXISTS groups.group_members (
-    group_id INTEGER REFERENCES groups.groups(id),
-    user_id INTEGER REFERENCES users.users(id),
+    group_id BIGINT REFERENCES groups.groups(id),
+    user_id BIGINT REFERENCES users.users(id),
     role VARCHAR(50) DEFAULT 'MEMBER',
     joined_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (group_id, user_id)
 );
 
 CREATE TABLE IF NOT EXISTS groups.group_content (
-    id SERIAL PRIMARY KEY,
-    group_id INTEGER REFERENCES groups.groups(id),
+    id BIGSERIAL PRIMARY KEY,
+    group_id BIGINT REFERENCES groups.groups(id),
     content_type VARCHAR(50) NOT NULL, -- flashcard_set, quiz, etc.
-    content_id INTEGER NOT NULL,
+    content_id BIGINT NOT NULL,
     added_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
